Update of file architecture refactoring
