+++ app/src/main/java/com/inercy/hidroponia/ui/components/BarProgress.kt
@@ -27,7 +27,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.inercy.hidroponia.R
-import com.inercy.hidroponia.ui.models.StatusEnum
+import com.inercy.hidroponia.ui.enums.StatusEnum
 
 /**
  * @param iconLabel Path of the drawable for the icon accompanying the label
