+++ app/src/main/java/com/inercy/hidroponia/ui/enums/StatusEnum.kt
@@ -1,4 +1,4 @@
-package com.inercy.hidroponia.ui.models
+package com.inercy.hidroponia.ui.enums
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckCircle
